project/requirements.txt
fastapi==0.62.0
asyncpg==0.20.1
uvicorn==0.12.3
tortoise-orm==0.16.12
pytest==6.1.2
requests==2.25.0
black==20.8b1